Cleaned up redundant code by making new functions to take care of duplicate code. Also moved parse_md into a new file called file_processors.py. Looking to make some performance imporvements in the next commit
