+++ txt_to_html.py
@@ -1,117 +1,7 @@
 import argparse
-import os
-from shutil import rmtree
-from helper import *
+from file_processors import text_to_html
 
 VERSION = "0.1"
-
-def remove_output_dir(
-        output_dir):  # function to remove output directory if it exists and make a new one regardless of if it exists or not
-    if os.path.exists(output_dir):
-        rmtree(output_dir)  # using rmtree to delete the directory even if it has files in it
-
-    os.makedirs(output_dir)  # Re/creating the output directory
-    
-def text_to_html(input_path, stylesheet, output_dir):
-    try:
-        if os.path.exists(input_path) and os.path.isdir(input_path):  # if the user inputted a directory
-            remove_output_dir(output_dir)
-            for filename in os.listdir(input_path):
-                if filename.endswith(".txt") or filename.endswith(".md"):
-
-                    input_file = os.path.join(input_path, filename)
-                    output_file = os.path.splitext(os.path.basename(input_file))[0] + ".html"  # Constructing the output file's path based on the name of the input file
-                    output_file = os.path.join(output_dir, output_file)
-                    print(output_file)
-                    
-                    if os.path.exists(output_file):
-                        output_file = generate_duplicate_filename(output_dir, output_file)
-
-                    # opening the input file
-                    with open(input_file, "r") as txt:
-                        lines = txt.readlines()
-
-                    html_contents = f"<html lang='en'>\n<head>\n\t<meta charset='utf-8'>\n"  # Creating the start of the html file
-
-                    title = os.path.splitext(os.path.basename(input_file))[0]  # Getting the file name and removing the path and txt extension so it can be used in the title tag
-                    html_contents += f"""\n \t<title>{title}</title>\n\t<meta name='viewport' content='width=device-width, initial-scale=1'>{f'<link rel="stylesheet" type="text/css" href="{stylesheet}">'}
-                    \n</head>\n<body>\n"""
-
-                    in_paragraph = False  # Flag to track if we are inside a paragraph
-                    for line in lines:
-                        converted_line = line
-
-                        if not in_paragraph:  # if in_paragraph is set to False, create the begining of the p tag in html_contnets and set in_paragraph to true
-                            html_contents += "<p>"
-                            in_paragraph = True
-
-                        if line == "\n":  # if the current line is a newline close the p tag and append it to then set in_paragraph to False
-                            html_contents += "</p>\n"
-                            in_paragraph = False
-
-                        else:  # if the code is still in a line with text and is in a paragraph append the line to html_contents
-                            converted_line = line.replace("\n", " ")
-                            html_contents += f" {converted_line}"
-
-                    # Closing the last paragraph if the loop was still in the paragraph when finished
-                    if in_paragraph:
-                        html_contents += "</p>\n"
-
-                    html_contents += f"</body>\n</html>"
-                    
-                    if filename.endswith(".md"):
-                        html_contents = parse_md(html_contents)
-
-                    with open(output_file, "w") as html:
-                        html.write(html_contents)
-
-            print("File conversion was successful! Please look for the ", output_dir, " folder.")
-
-        elif (input_path.endswith(".txt") or input_path.endswith(".md")) and os.path.isfile(input_path):
-            remove_output_dir(output_dir)
-
-            output_file = os.path.splitext(os.path.basename(input_path))[0] + ".html" # Constructing the output file's path based on the name of the input file
-            output_file = os.path.join(output_dir, output_file)
-            input_file = input_path
-
-            with open(input_file, "r") as txt:
-                lines = txt.readlines()
-
-            html_contents = f"<html lang='en'>\n<head>\n\t<meta charset='utf-8'>\n"  # Creating the start of the html file
-            title = os.path.splitext(os.path.basename(input_file))[0]  # Creating the title based off of the filename
-            html_contents += f"""\n \t<title>{title}</title>\n\t<meta name='viewport' content='width=device-width, initial-scale=1'>{f'<link rel="stylesheet" type="text/css" href="{stylesheet}">'}
-                                \n</head>\n<body>\n"""
-
-            in_paragraph = False  # Flag to track if we are inside a paragraph
-            for line in lines:
-                converted_line = line
-                if not in_paragraph:  # if in_paragraph is set to False, create the begining of the p tag in html_contnets and set in_paragraph to true
-                    html_contents += "<p>"
-                    in_paragraph = True
-                if line == "\n":  # if the current line is a newline close the p tag and append it to then set in_paragraph to False
-                    html_contents += "</p>\n"
-                    in_paragraph = False
-                else:  # if the code is still in a line with text and is in a paragraph append the line to html_contents
-                    converted_line = line.replace("\n", "")
-                    html_contents += f" {converted_line}"
-            # Closing the last paragraph if the loop was still in the paragraph when finished
-            if in_paragraph:
-                html_contents += "</p>\n"
-
-            html_contents += f"</body>\n</html>"
-            
-            if input_path.endswith(".md"):
-                html_contents = parse_md(html_contents)
-
-            with open(output_file, "w") as html:
-                html.write(html_contents)
-            print("File conversion was successful! Please look for the ", output_dir, " folder.")
-
-        else:  # if the file/folder entered does not exist print error message
-            print("The file/directory name does not exist. Please make sure you entered the correct name")
-
-    except Exception as e:  # Throw an error if any kind of error happens
-        print(f"An error occurred: {str(e)}")
 
 
 if __name__ == '__main__':
